pyproject.toml
[project]
name = "wildlifemapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "pycocotools>=2.0.10",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.6",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.2",
    "scipy>=1.16.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
